BaseODE superclass created with basic implementation
