+++ aviary/interface/default_phase_info/height_energy_fiti.py
@@ -45,7 +45,8 @@ phase_info = {
 }
 
 
-def phase_info_parameterization(phase_info, post_mission_info, aviary_inputs: AviaryValues):
+def phase_info_parameterization(phase_info, post_mission_info,
+                                aviary_inputs: AviaryValues):
     """
     Modify the values in the phase_info dictionary to accomodate different values
     for the following mission design inputs: cruise altitude, cruise mach number,
