+++ aviary/mission/gasp_based/ode/landing_ode.py
@@ -2,7 +2,7 @@ import numpy as np
 
 from aviary.subsystems.atmosphere.atmosphere import Atmosphere
 
-from aviary.mission.gasp_based.ode.base_ode import BaseODE
+from aviary.mission.gasp_based.ode.two_dof_ode import BaseODE
 from aviary.mission.gasp_based.ode.params import ParamPort
 from aviary.mission.gasp_based.ode.landing_eom import (
     GlideConditionComponent,
