+++ aviary/mission/gasp_based/ode/rotation_ode.py
@@ -1,7 +1,7 @@
 import numpy as np
 import openmdao.api as om
 
-from aviary.mission.gasp_based.ode.base_ode import BaseODE
+from aviary.mission.gasp_based.ode.two_dof_ode import BaseODE
 from aviary.mission.gasp_based.ode.params import ParamPort
 from aviary.mission.gasp_based.ode.rotation_eom import RotationEOM
 from aviary.variable_info.enums import AnalysisScheme
