+++ aviary/mission/gasp_based/ode/taxi_ode.py
@@ -6,7 +6,7 @@ from aviary.utils.aviary_values import AviaryValues
 from aviary.utils.functions import add_opts2vals, create_opts2vals
 
 from aviary.variable_info.enums import SpeedType
-from aviary.mission.gasp_based.ode.base_ode import BaseODE
+from aviary.mission.gasp_based.ode.two_dof_ode import BaseODE
 from aviary.mission.gasp_based.ode.params import ParamPort
 from aviary.mission.gasp_based.ode.taxi_eom import TaxiFuelComponent
 from aviary.subsystems.atmosphere.atmosphere import Atmosphere
